Decouple configuration from deployment artifact (#45)
Allow overriding settings with environment variables.

Leveraging the environment variables in AWS Lambda makes it possible to
include the bless_deploy.cfg in the same repo, without exposing secrets
or to deploy the same zip with multiple configurations


When deploying the same zip in mulitple regions, you can leave out the
region_passwprd option and set the default_password option with environment
variables. This allows you the change the same variable in every region

* Add 'ca_private_key' option

This extra option allows passing in the (encrypted) private key directly.
When setting this with an environment variable it can be used to have one
zip that can be deployed with different ca's.